REFACTOR: change file/directory name and fix css error
