+++ src/pages/main/InfoListTemplate.tsx
@@ -8,6 +8,9 @@ const InfoListTemplate: FC = ({ children }) => {
 const Container = styled.div`
   max-width: 1130px;
   margin: 30px auto;
+  @media (min-width: 600px) {
+    min-width: 1130px;
+  }
 `;
 
 export default InfoListTemplate;
